NeighborlyAPI/requirements.txt
azure-functions==1.17.0
dnspython==2.4.2
pymongo==4.6.1
